Move notifications_row file out of sidebar, fix subsequent linking.
